split up publish code that can be reused
