+++ packages/xdl/src/project/publishAsync.ts
@@ -2,12 +2,9 @@ import {
   ExpoAppManifest,
   ExpoConfig,
   getDefaultTarget,
+  HookArguments,
   PackageJSONConfig,
-  Platform,
 } from '@expo/config';
-import { bundleAsync } from '@expo/dev-server';
-import axios from 'axios';
-import chalk from 'chalk';
 import FormData from 'form-data';
 import fs from 'fs-extra';
 import path from 'path';
@@ -17,183 +14,16 @@ import ApiV2 from '../ApiV2';
 import Config from '../Config';
 import * as EmbeddedAssets from '../EmbeddedAssets';
 import { shouldUseDevServer } from '../Env';
-import { ErrorCode } from '../ErrorCode';
 import logger from '../Logger';
 import { publishAssetsAsync } from '../ProjectAssets';
 import * as Sentry from '../Sentry';
-import * as UrlUtils from '../UrlUtils';
 import UserManager, { User } from '../User';
 import XDLError from '../XDLError';
 import * as ExponentTools from '../detach/ExponentTools';
-import * as TableText from '../logs/TableText';
-import { learnMore } from '../logs/TerminalLink';
-import {
-  startReactNativeServerAsync,
-  stopReactNativeServerAsync,
-} from '../start/startLegacyReactNativeServerAsync';
-import { resolveEntryPoint } from '../tools/resolveEntryPoint';
 import * as Doctor from './Doctor';
-import * as ProjectUtils from './ProjectUtils';
+import { buildPublishBundlesAsync, printBundleSizes } from './buildPublishBundlesAsync';
 import { getPublishExpConfigAsync, PublishOptions } from './getPublishExpConfigAsync';
 import { LoadedHook, prepareHooks, runHook } from './runHook';
-
-const MINIMUM_BUNDLE_SIZE = 500;
-
-type PackagerOptions = {
-  dev: boolean;
-  minify: boolean;
-};
-
-export async function buildPublishBundlesAsync(
-  projectRoot: string,
-  publishOptions: PublishOptions = {},
-  bundleOptions: { dev?: boolean; useDevServer: boolean }
-) {
-  if (!bundleOptions.useDevServer) {
-    try {
-      await startReactNativeServerAsync({
-        projectRoot,
-        options: {
-          nonPersistent: true,
-          maxWorkers: publishOptions.maxWorkers,
-          target: publishOptions.target,
-          reset: publishOptions.resetCache,
-        },
-        verbose: !publishOptions.quiet,
-      });
-      return await fetchPublishBundlesAsync(projectRoot);
-    } finally {
-      await stopReactNativeServerAsync(projectRoot);
-    }
-  }
-
-  const platforms: Platform[] = ['android', 'ios'];
-  const [android, ios] = await bundleAsync(
-    projectRoot,
-    {
-      target: publishOptions.target,
-      resetCache: publishOptions.resetCache,
-      logger: ProjectUtils.getLogger(projectRoot),
-      quiet: publishOptions.quiet,
-    },
-    platforms.map((platform: Platform) => ({
-      platform,
-      entryPoint: resolveEntryPoint(projectRoot, platform),
-      dev: bundleOptions.dev,
-    }))
-  );
-
-  return {
-    android,
-    ios,
-  };
-}
-
-// Fetch iOS and Android bundles for publishing
-async function fetchPublishBundlesAsync(projectRoot: string, opts?: PackagerOptions) {
-  const entryPoint = resolveEntryPoint(projectRoot);
-  const publishUrl = await UrlUtils.constructPublishUrlAsync(
-    projectRoot,
-    entryPoint,
-    undefined,
-    opts
-  );
-  const sourceMapUrl = await UrlUtils.constructSourceMapUrlAsync(projectRoot, entryPoint);
-  const assetsUrl = await UrlUtils.constructAssetsUrlAsync(projectRoot, entryPoint);
-
-  logger.global.info('Building iOS bundle');
-  const iosBundle = await _getForPlatformAsync(projectRoot, publishUrl, 'ios', {
-    errorCode: 'INVALID_BUNDLE',
-    minLength: MINIMUM_BUNDLE_SIZE,
-  });
-
-  logger.global.info('Building Android bundle');
-  const androidBundle = await _getForPlatformAsync(projectRoot, publishUrl, 'android', {
-    errorCode: 'INVALID_BUNDLE',
-    minLength: MINIMUM_BUNDLE_SIZE,
-  });
-
-  logger.global.info('Building source maps');
-  const iosSourceMap = await _getForPlatformAsync(projectRoot, sourceMapUrl, 'ios', {
-    errorCode: 'INVALID_BUNDLE',
-    minLength: MINIMUM_BUNDLE_SIZE,
-  });
-  const androidSourceMap = await _getForPlatformAsync(projectRoot, sourceMapUrl, 'android', {
-    errorCode: 'INVALID_BUNDLE',
-    minLength: MINIMUM_BUNDLE_SIZE,
-  });
-
-  logger.global.info('Building asset maps');
-  const iosAssetsJson = await _getForPlatformAsync(projectRoot, assetsUrl, 'ios', {
-    errorCode: 'INVALID_ASSETS',
-  });
-  const androidAssetsJson = await _getForPlatformAsync(projectRoot, assetsUrl, 'android', {
-    errorCode: 'INVALID_ASSETS',
-  });
-
-  return {
-    android: { code: androidBundle, map: androidSourceMap, assets: JSON.parse(androidAssetsJson) },
-    ios: { code: iosBundle, map: iosSourceMap, assets: JSON.parse(iosAssetsJson) },
-  };
-}
-
-async function _getForPlatformAsync(
-  projectRoot: string,
-  url: string,
-  platform: Platform,
-  { errorCode, minLength }: { errorCode: ErrorCode; minLength?: number }
-): Promise<string> {
-  const fullUrl = `${url}&platform=${platform}`;
-  let response;
-
-  try {
-    response = await axios.request({
-      url: fullUrl,
-      responseType: 'text',
-      // Workaround for https://github.com/axios/axios/issues/907.
-      // Without transformResponse, axios will parse the body as JSON regardless of the responseType/
-      transformResponse: [data => data],
-      proxy: false,
-      validateStatus: status => status === 200,
-      headers: {
-        'Exponent-Platform': platform,
-      },
-    });
-  } catch (error) {
-    if (error.response) {
-      if (error.response.data) {
-        let body;
-        try {
-          body = JSON.parse(error.response.data);
-        } catch (e) {
-          ProjectUtils.logError(projectRoot, 'expo', error.response.data);
-        }
-
-        if (body) {
-          if (body.message) {
-            ProjectUtils.logError(projectRoot, 'expo', body.message);
-          } else {
-            ProjectUtils.logError(projectRoot, 'expo', error.response.data);
-          }
-        }
-      }
-      throw new XDLError(
-        errorCode,
-        `Packager URL ${fullUrl} returned unexpected code ${error.response.status}. ` +
-          'Please open your project in the Expo app and see if there are any errors. ' +
-          'Also scroll up and make sure there were no errors or warnings when opening your project.'
-      );
-    } else {
-      throw error;
-    }
-  }
-
-  if (!response.data || (minLength && response.data.length < minLength)) {
-    throw new XDLError(errorCode, `Body is: ${response.data}`);
-  }
-
-  return response.data;
-}
 
 export interface PublishedProjectResult {
   /**
@@ -244,37 +74,18 @@ export async function publishAsync(
   const bundles = await buildPublishBundlesAsync(projectRoot, options, {
     useDevServer: shouldUseDevServer(exp),
   });
-  const androidBundle = bundles.android.code;
-  const iosBundle = bundles.ios.code;
 
-  const files = [
-    ['index.ios.js', bundles.ios.code],
-    ['index.android.js', bundles.android.code],
-  ];
-  // Account for inline source maps
-  if (bundles.ios.map) {
-    files.push([chalk.dim('index.ios.js.map'), bundles.ios.map]);
-  }
-  if (bundles.android.map) {
-    files.push([chalk.dim('index.android.js.map'), bundles.android.map]);
-  }
-
-  logger.global.info('');
-  logger.global.info(TableText.createFilesTable(files));
-  logger.global.info('');
-  logger.global.info(
-    `💡 JavaScript bundle sizes affect startup time. ${chalk.dim(
-      learnMore(`https://expo.fyi/javascript-bundle-sizes`)
-    )}`
-  );
-  logger.global.info('');
+  printBundleSizes(bundles);
 
   await publishAssetsAsync({ projectRoot, exp, bundles });
 
+  const androidBundle = bundles.android.code;
+  const iosBundle = bundles.ios.code;
+
   const hasHooks = validPostPublishHooks.length > 0;
 
-  const shouldPublishAndroidMaps = !!exp.android && !!exp.android.publishSourceMapPath;
-  const shouldPublishIosMaps = !!exp.ios && !!exp.ios.publishSourceMapPath;
+  const shouldPublishAndroidMaps = !!exp.android?.publishSourceMapPath;
+  const shouldPublishIosMaps = !!exp.ios?.publishSourceMapPath;
   const androidSourceMap = hasHooks || shouldPublishAndroidMaps ? bundles.android.map : null;
   const iosSourceMap = hasHooks || shouldPublishIosMaps ? bundles.ios.map : null;
 
@@ -302,8 +113,8 @@ export async function publishAsync(
 
   if (
     validPostPublishHooks.length ||
-    (exp.ios && exp.ios.publishManifestPath) ||
-    (exp.android && exp.android.publishManifestPath) ||
+    exp.ios?.publishManifestPath ||
+    exp.android?.publishManifestPath ||
     EmbeddedAssets.shouldEmbedAssetsForExpoUpdates(projectRoot, exp, pkg, target)
   ) {
     [androidManifest, iosManifest] = await Promise.all([
@@ -321,7 +132,7 @@ export async function publishAsync(
       }),
     ]);
 
-    const hookOptions = {
+    const hookOptions: Omit<HookArguments, 'config'> = {
       url: response.url,
       exp,
       iosBundle,
